Silence compiler warnings

+++ src/xbee.c
@@ -230,9 +230,9 @@ static void xbeeStatsAdd(struct XBeeStaticticsSummary *summary, struct timeval c
 
 static void xbeeStatsSummarise(struct XBeeStaticticsSummary const *summary) {
   pmsg_notice("  Minimum response time: %lu.%06lu\n",
-    summary->minimum.tv_sec, (unsigned long) summary->minimum.tv_usec);
+    (unsigned long) summary->minimum.tv_sec, (unsigned long) summary->minimum.tv_usec);
   pmsg_notice("  Maximum response time: %lu.%06lu\n",
-    summary->maximum.tv_sec, (unsigned long) summary->maximum.tv_usec);
+    (unsigned long) summary->maximum.tv_sec, (unsigned long) summary->maximum.tv_usec);
 
   struct timeval average;
 
@@ -247,7 +247,8 @@ static void xbeeStatsSummarise(struct XBeeStaticticsSummary const *summary) {
   average.tv_sec += usecs/1000000;
   average.tv_usec = usecs%1000000;
 
-  pmsg_notice("  Average response time: %lu.%06lu\n", average.tv_sec, (unsigned long) average.tv_usec);
+  pmsg_notice("  Average response time: %lu.%06lu\n", (unsigned long) average.tv_sec,
+    (unsigned long) average.tv_usec);
 }
 
 static void XBeeBootSessionInit(struct XBeeBootSession *xbs) {
